Changes on Login, Dashboard
