+++ src/layouts/SidebarLayout/Header/index.tsx
@@ -1,5 +1,6 @@
 import {
   Box,
+  Button,
   Divider,
   Stack,
   alpha,
@@ -9,6 +10,9 @@ import {
 } from '@mui/material';
 
 import HeaderUserbox from './Userbox';
+import { useRouter } from 'next/router';
+import Cookies from "js-cookie";
+import { useEffect, useState } from 'react';
 
 const HeaderWrapper = styled(Box)(
   ({ theme }) => `
@@ -31,6 +35,20 @@ const HeaderWrapper = styled(Box)(
 
 function Header() {
   const theme = useTheme();
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const router = useRouter();
+  const { query } = router;
+  const userName = query.name || '';
+
+  useEffect(() => {
+    const token = Cookies.get('token_person');
+    const role = Cookies.get('role');
+    if (token && role) {
+      setIsAuthenticated(true);
+    } else {
+      setIsAuthenticated(false);
+    }
+  }, []);
 
   return (
     <HeaderWrapper
@@ -40,16 +58,16 @@ function Header() {
         boxShadow:
           theme.palette.mode === 'dark'
             ? `0 1px 0 ${alpha(
-                lighten(theme.colors.primary.main, 0.7),
-                0.15
-              )}, 0px 2px 8px -3px rgba(0, 0, 0, 0.2), 0px 5px 22px -4px rgba(0, 0, 0, .1)`
+              lighten(theme.colors.primary.main, 0.7),
+              0.15
+            )}, 0px 2px 8px -3px rgba(0, 0, 0, 0.2), 0px 5px 22px -4px rgba(0, 0, 0, .1)`
             : `0px 2px 8px -3px ${alpha(
-                theme.colors.alpha.black[100],
-                0.2
-              )}, 0px 5px 22px -4px ${alpha(
-                theme.colors.alpha.black[100],
-                0.1
-              )}`
+              theme.colors.alpha.black[100],
+              0.2
+            )}, 0px 5px 22px -4px ${alpha(
+              theme.colors.alpha.black[100],
+              0.1
+            )}`
       }}
     >
       <Stack
@@ -60,7 +78,18 @@ function Header() {
       >
       </Stack>
       <Box display="flex" alignItems="center">
-        <HeaderUserbox />
+        {!isAuthenticated ? (
+          <>
+            <Button href="/login" variant="outlined">
+              Iniciar sesion
+            </Button>
+            <Button href="/singup" sx={{ margin: 1 }} color="secondary">
+              Registrarse
+            </Button>
+          </>
+        ) : (
+          <HeaderUserbox userName={userName} />
+        )}
       </Box>
     </HeaderWrapper>
   );
