+++ src/layouts/SidebarLayout/Sidebar/index.tsx
@@ -53,14 +53,14 @@ function Sidebar() {
         }}
       >
         <Scrollbar>
-          <Box mt={3}>
+          <Box mt={2} height={55}>
             <Box
-              mx={2}
+              mx={3}
               sx={{
-                width: 52
+                 margin: 'auto'
               }}
             >
-              <Logo />
+              <Logo></Logo>
             </Box>
           </Box>
           <Divider
