+++ src/layouts/SidebarLayout/index.tsx
@@ -1,6 +1,5 @@
 import { Box, alpha, lighten, useTheme } from '@mui/material';
 import { FC, ReactNode } from 'react';
-
 import Header from './Header';
 import PropTypes from 'prop-types';
 import Sidebar from './Sidebar';
